# EVM: Initialize memory with CONTAINER_SIZE bytes