This is a small performance optimization improving `MLOAD` and `MSTORE` performance by 10-20%.

Performance results (always last block):

cd ../evm && npm run build && cd ../client && npm run client:start -- --sync=none --vmProfileBlocks --executeBlocks=962960-962971 (MLOAD 48, MSTORE 100 times)
MLOAD
New 2.8  2.7 2.8 2.8
Old. 2.65 2.8 2.7 2.5
MSTORE
New 1.85 1.8 1.9 1.8
Old. 1.65  1.8 1.7 1.75

cd ../evm && npm run build && cd ../client && npm run client:start -- --sync=none --vmProfileBlocks --executeBlocks=962960-962970 (MLOAD 1224, MSTORE 2550 times)
MLOAD
New 2.8 2.75 2.7 2.65
Old.  2.85 2.8 2.7 2.55
MSTORE
New 1.85 1.85 1.8 1.7
Old. 1.75. 1.7 1.7 1.65

Ok. So I would think that results are clear enough on this and we can take this in. Ready for review. 🙂 
We should nevertheless really do this normalized MGas/s (as discussed with @jochem-brouwer), still think this would make comparison significantly easier (high level idea: normalize by constant factor to factor out runtime differences).

Might take this on in a separate PR.

Think this should nevertheless be fine.