@holgerd77 2023-09-18
@jochem-brouwer Thanks for approval will merge! 👍 

> However, your description does not seem to match, you seem to run the exact same test (except the second test is one block less than the first one) - but the amount of MSTORE/MLOADS differ.

TBH I cannot follow yet what you mean here (there are only two tests?).