@jochem-brouwer 2023-09-18
You run:

`cd ../evm && npm run build && cd ../client && npm run client:start -- --sync=none --vmProfileBlocks --executeBlocks=962960-962971 (MLOAD 48, MSTORE 100 times)`

and

`cd ../evm && npm run build && cd ../client && npm run client:start -- --sync=none --vmProfileBlocks --executeBlocks=962960-962970 (MLOAD 1224, MSTORE 2550 times)`

(per OP)

These are almost the same except the second test runs until 962970 (so one block less), yet is has somehow more MLOADs and MSTOREs? :thinking: