@holgerd77 2023-09-18
Results (and MLOAD / MSTORE counts) are always only for the last block. I'm always testing like this, running some blocks before, to avoid interference with client startup stuff.