@jochem-brouwer 2023-09-18
Oh right, this makes a lot of sense (to warmup the client). Thanks for clarifying :smile: